iron/requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
canvasapi>=3.0.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dateutil>=2.8.2
pytz>=2023.3
aioimaplib>=1.0.1
mcstatus>=11.0.0
Pillow>=10.0.0
dateparser>=1.1.8
